Security

+++ src/AppBundle/Controller/TaskController.php
@@ -86,10 +86,28 @@ class TaskController extends Controller
     public function deleteTaskAction(Task $task)
     {
         $em = $this->getDoctrine()->getManager();
-        $em->remove($task);
-        $em->flush();
 
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
+        $author = $task->getAuthor();
+
+        if ($author == null) {
+            if ($this->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+                $em->remove($task);
+                $em->flush();
+
+                $this->addFlash('success', 'La tâche a bien été supprimée.');
+            } else {
+                $this->addFlash('error', 'Vous devez être administrateur pour supprimer une tâche anonyme.');
+            }
+        } else {
+            if ($author == $this->get('security.token_storage')->getToken()->getUser()) {
+                $em->remove($task);
+                $em->flush();
+
+                $this->addFlash('success', 'La tâche a bien été supprimée.');
+            } else {
+                $this->addFlash('error', 'Vous ne pouvez supprimer que vos propres tâches.');
+            }
+        }
 
         return $this->redirectToRoute('task_list');
     }
